Make dependencies more inline with github pages
This seems to fix a rendering bug with some markdown I’m about to
commit.

Note: we could just use the github-provided gem to keep all our deps in
sync with gh-pages, but in case we ever move to something else, I
figure we might as well keep listing the dependencies explicitly.