Error objects: specify `missing` member
This will be used for a server to indicate which profile extension is
missing from a client’s document, in the event that a server requires
an extension be used (e.g. for authentication).

Note: if the markdown formatting looks weird…it is, but that’s the only
formatting that would put the “This object…” paragraph under the
`missing` bullet.